% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcape.R
\name{cape_cin}
\alias{cape_cin}
\title{Compute CAPE/CIN according to George H. Bryan formula}
\usage{
cape_cin(
  p_in,
  t_in,
  td_in,
  pinc = 100,
  source = 1,
  ml_depth = 200,
  adiabat = 4
)
}
\arguments{
\item{pinc}{= 2.0 # Pressure increment (Pa)}

\item{source}{Source parcel: 1 = surface, 2 = most unstable (max theta-e), 3 = mixed-layer (specify ml_depth)}

\item{ml_depth}{depth (m) of mixed layer (for source = 3); default 200}

\item{adiabat}{= 4   #! Formulation of moist adiabat: 1 = pseudoadiabatic, liquid only, 2 = reversible, liquid only, 3 = pseudoadiabatic, with ice,  4 = reversible, with ice}
}
\description{
TODO
To be completed..
}
\examples{
p_in = c(1000, 855, 700, 500, 300, 100, 10)
t_in = c(25, 10, 0, -15, -30, -50, -92)
td_in = c(20, 5, -5, -30, -55, -80, -99)


}
